Place order functions updated

+++ Backend/src/main/java/lk/ijse/gdse66/springboot/shoeshopmanagementsystem/backend/entity/Sale.java
@@ -20,7 +20,7 @@ public class Sale {
     private Double totalPrice;
     private Timestamp purchaseDate;
     private String paymentMethod;
-    private Double addedPoints;
+    private Integer addedPoints;
     private String cashierName;
 
     @ManyToOne(fetch = FetchType.LAZY)
